requirements.txt
Flask==1.1.2
Flask-Login==0.4.1
oauthlib==3.0.1
pyOpenSSL==19.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coverage==5.4
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
PyQt5==5.15.2
PyQt5-sip==12.8.1
requests==2.25.1
urllib3==1.26.5
Werkzeug==1.0.1
psycopg2-binary==2.8.6
pip==21.0.1
pytz==2021.1
